+++ pages/components/lib/Utils.js
@@ -1,6 +1,6 @@
 import { clsx } from "clsx";
 import { twMerge } from "tailwind-merge";
 
-export function cn(...inputs) {
+export default function cn(...inputs) {
   return twMerge(clsx(inputs));
 }
