--- pages/components/timeline/Timeline.js
@@ -1,58 +1,0 @@
-export const dataTimeline = [
-  {
-    title: "Preencha o formulário",
-    content: (
-      <div>
-        <p className="text-neutral-800 dark:text-neutral-200 text-xs md:text-sm font-normal mb-8">
-          Dê o primeiro passo e apresente sua empresa. Vamos ajudar você a
-          alcançar novos patamares.
-        </p>
-        <div className="text-neutral-800 dark:text-neutral-200">
-          📝 Dados seguros e protegidos
-        </div>
-      </div>
-    ),
-  },
-  {
-    title: "Conversa com especialistas",
-    content: (
-      <div>
-        <p className="text-neutral-800 dark:text-neutral-200 text-xs md:text-sm font-normal mb-8">
-          Em até 12 horas, nossa equipe entra em contato para entender seu
-          negócio e suas metas.
-        </p>
-        <div className="text-neutral-800 dark:text-neutral-200">
-          📅 Marque sua agenda
-        </div>
-      </div>
-    ),
-  },
-  {
-    title: "Planejamento sob medida",
-    content: (
-      <div>
-        <p className="text-neutral-800 dark:text-neutral-200 text-xs md:text-sm font-normal mb-8">
-          Receba uma proposta estratégica com foco em resultados e
-          fortalecimento de sua marca.
-        </p>
-        <div className="text-neutral-800 dark:text-neutral-200">
-          🔍 Análise personalizada
-        </div>
-      </div>
-    ),
-  },
-  {
-    title: "Acompanhe o crescimento",
-    content: (
-      <div>
-        <p className="text-neutral-800 dark:text-neutral-200 text-xs md:text-sm font-normal mb-8">
-          Implementamos e monitoramos o sucesso de sua marca, ajustando
-          estratégias para alcançar os melhores resultados.
-        </p>
-        <div className="text-neutral-800 dark:text-neutral-200">
-          🚀 Alavanque sua marca
-        </div>
-      </div>
-    ),
-  },
-];
