+++ pages/components/ui/Google-gemini-effect.js
@@ -162,3 +162,5 @@ export const GoogleGeminiEffect = ({
     </div>
   );
 };
+
+export default GoogleGeminiEffect;
