+++ pages/components/ui/Moving-border.js
@@ -121,3 +121,5 @@ export const MovingBorder = ({
     </>
   );
 };
+
+export default Button;
