--- pages/inicio/inicio.module.css
@@ -1,3 +1,0 @@
-#hero {
-  background-size: contain;
-}
